requirements.txt
Flask
pymongo
python-dotenv
qrcode[pil]
mongomock
